refactor(tests): split role registry tests from test_governance.py
Move TestBuildLabelDispatchEvent and TestDispatchPredicates to
dedicated test_role_registry.py for better maintainability.

Before:
- test_governance.py: 695 lines, 10 classes, 37 tests

After:
- test_governance.py: 609 lines, 8 classes, 30 tests (core governance)
- test_role_registry.py: 99 lines, 2 classes, 7 tests (dispatch helpers)

This is NOT a revert to the original 3-file structure:
- Original: test_governance.py + test_governance_context.py + test_registry.py
- Now: test_governance.py + test_role_registry.py (2 files, cleaner split)

Closes #1336
